Changed more names
This is probably a better representation of what I’ve been doing.

All that we need to do to display a JSON/bodymovin animation file is
link it into the “doAnimation()” function (as well as have a DOM
element by the name of “home” in our HTML file)

The next step is to have the “doAnimation()” function add play/pause
buttons onto the div that it creates. After that, I should feel
comfortable with messing with the bodymovin side of this (as opposed to
the JavaScript side of this).

Then, I’ll port it all to React, and change our “handle” on this whole
process from a “doAnimation()” function to a HOC constructor call or
something like that.